+++ GUI/WinForge.GUI/Services/Implementations/VersionServiceImpl.cs
@@ -16,6 +16,7 @@
 
 using System.Diagnostics;
 using System.IO;
+using System.Text;
 using System.Text.Json;
 using System.Text.RegularExpressions;
 using WinForge.GUI.Services;
@@ -106,6 +107,8 @@ public partial class VersionServiceImpl : IVersionService
                 UseShellExecute = false,
                 RedirectStandardOutput = true,
                 RedirectStandardError = true,
+                StandardOutputEncoding = Encoding.UTF8,
+                StandardErrorEncoding = Encoding.UTF8,
                 CreateNoWindow = true
             };
 
@@ -160,6 +163,8 @@ public partial class VersionServiceImpl : IVersionService
                 UseShellExecute = false,
                 RedirectStandardOutput = true,
                 RedirectStandardError = true,
+                StandardOutputEncoding = Encoding.UTF8,
+                StandardErrorEncoding = Encoding.UTF8,
                 CreateNoWindow = true
             };